Refactor variables page into a layout
Removes some prop drilling
